Move PersonnelReport methods
